Update the interactive tag when necessary
Update the interactive tag on entity metadata update, item slot change, and sneaking update; that way, if the tag should change, it will change.
